changed pages structure and moved YMaps provider
