--- src/components/Service/Service.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Service = () => {
-    return (
-        <div>
-
-        </div>
-    );
-};
-
-export default Service;
