+++ src/pages/Trainings/Trainings.module.scss
