+++ src/router.js
@@ -1,10 +1,10 @@
 import {createBrowserRouter} from "react-router-dom";
 import {Main} from "./components/Layout/Main/Main/Main";
-import {Contacts} from "./components/Contacts/Contacts";
+import {Contacts} from "./pages/Contacts/Contacts";
 import {About} from "./components/Layout/About";
 import {OnlineRecording} from "./components/Layout/OnlineRecording";
-import {Services} from "./components/Layout/Services";
-import {Trainings} from "./components/Trainings/Trainings";
+import {Services} from "./pages/Services/Services";
+import {Trainings} from "./pages/Trainings/Trainings";
 import {Login} from "./pages/Login";
 import {Registration} from "./pages/Registration";
 import {Specialists} from "./components/Layout/Specialists/Specialists";
